+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Fixed
 - [RT#211209]: reallocation of a previously allocated doi.
+- Resetting the position of the element when it is removed from this volume does not work.
 ## v1.0.43 - 2024-04-17
 ### Changed
 - Updated application's requirements to PHP 8.1.* . All application dependencies were updated accordingly